refactor: break the util <-> computeengine import cycle
`util` is a utility layer that reached back up into the core engine from
inside seven function bodies, so the dependency edge was real while being
invisible to a module-level reader: `C`, `ConstantValue`,
`identity_function` and `_bind_self` are unconditional core vocabulary
that `util` needs on every call, deferred only to survive the cycle that
`computeengine`'s module-level `from .util import ...` closes. A second
cycle rode on the same fault, with `nodekey` importing `as_iterable`
from `util` at module level while `util` deferred its own import of
`nodekey.to_nodekey`.

Two leaf modules now hold what both sides need:

- `loman.nodedefs` --- `ConstantValue`/`C`, `identity_function` and
  `_bind_self`. Imports nothing from loman but `compat`, itself a leaf.
- `loman.iterables` --- `apply1`, `as_iterable` and `apply_n`. Imports
  nothing from loman at all, which is what lets `nodekey` and
  `graph_utils` stop depending on the whole utility layer.

Every deferred loman import in `util.py` is gone, the four hoisted
`loman.consts` imports included: `consts` is a leaf and those edges were
invisible for no reason. `util` now imports `consts`, `iterables`,
`nodedefs` and `nodekey` at module level, and names `Computation` under
`TYPE_CHECKING` only --- the one edge that cannot be a runtime import,
since the engine imports `util`.

`_bind_self` had no caller in `computeengine`; `util` is its only one, so
it moves rather than being re-exported. The three names the engine does
use are imported back from `nodedefs`, so `loman.C` and
`loman.computeengine.ConstantValue` still resolve. Internally moved:
`loman.util.as_iterable` and `loman.util.apply_n` (now `loman.iterables`)
and `loman.computeengine._bind_self` (now `loman.nodedefs`).

Verified: an AST pass over `src/loman` reports no runtime module-level
cycle anywhere in the package and no deferred loman import left in
`util.py` or `nodekey.py`. 1541 tests pass, coverage 98.94% with both new
modules at 100%, `ty check src` reports the same single pre-existing
warning as master, and every pre-commit hook passes.

Not addressed: `computeengine.py` is still rank C (0.00) at 2634 lines.
The cycle was the structural reason anything `util` needed had to live
there; moving code out of it is now possible but is separate work.

Closes #391.
