+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [unreleased]
+- CHORE: Broke the import cycle between `loman.computeengine` and `loman.util`. The utility layer reached back into the engine from inside seven function bodies --- `C`, `ConstantValue`, `identity_function` and `_bind_self` --- because importing them at module level would have closed a loop, and a second cycle ran through `loman.nodekey`, which imported `as_iterable` from `util` while `util` deferred its own import of `to_nodekey`. That shared vocabulary now lives in two leaf modules, `loman.nodedefs` and `loman.iterables`, so every dependency `util` has is visible at the top of the file and a symbol the engine happens to define is no longer trapped there. No name in `loman`'s public surface moved; the internal ones that did are `loman.util.as_iterable` and `loman.util.apply_n`, now in `loman.iterables`, and `loman.computeengine._bind_self`, now in `loman.nodedefs`
 - FEAT: The widget can now build the graph, not only follow it. `comp.widget(buildable=True)` adds a node form to the toolbar and an Edit/Rename/Delete section to the detail panel, so nodes can be added, redefined, renamed and deleted without leaving the picture. An input node is a name and optionally a scalar; a calculation node is a name, a list of inputs and a Python expression, where each input becomes both a parameter of the compiled function and an edge in the graph. Everything maps onto the API that already existed --- `add_node`, `rename_node`, `delete_node` --- so a node built in the widget is not a second class of node
 - FEAT: Names in the node form are read against the block in focus, so a name typed while inside `market` lands inside `market`, and a leading `/` names a node from the top of the computation. That is what lets a node in a block depend on one outside it, and it is why there is no "new block" button: a block is a naming convention, so adding `market/spot` and `market/vol` makes the block appear around them
 - FEAT: Building the graph is opt-in through `buildable=True`, off by default and separate from `editable`, because defining a calculation node compiles an expression written in the browser and runs it in the kernel. `namespace=globals()` is what puts the notebook's own imports in scope for it; without a namespace only builtins are. The function's globals stay pointed at the live mapping, so an import made after a node was built is visible to it, and defining a node does not bind its name in the caller's namespace
